+++ src/index.ts
@@ -1,4 +1,5 @@
 export {Vector} from './Vector.js';
-export {Vector2D} from './2D.js';
+export {Vector2D, PolarVector2D} from './2D.js';
 export {Vector3D} from './3D.js';
 export {Complex} from './Complex.js';
+export {GeneralVector} from './General.js';